Require x11-font-daewoo-misc, x11-font-isas-misc and x11-font-jis-misc
If the fonts required by libxt are not present, xlib can spend a lot of time
looking for them, which makes simple apps like "xmessage" and "xcalc" take
almost 10 seconds to start on systems with a lot of fonts. This case is usually
triggered by Xt apps that call XtSetLanguageProc on UTF-8 systems. Since UTF-8
is the default, require the fonts.
CCBUG: 60967


SILENT: svn-revision: 603126